added header and footer code provided by jvanpelt
Header

Please place this script inside of the <head> tags to ensure that the tracking script loads before the rest of the page loads. (I did inquire about using defer or async, but we’ve been asked to stick strictly to Adobe’s guidelines on this one.)

<script type="text/javascript" src="//www.redhat.com/dtm-staging.js"></script>

Footer

This second tag needs to be placed in the body just before the close body tag </body>. We want to ensure that the entire page has loaded before we count it as a page load in our metrics.

<script type="text/javascript">
if (("undefined" !== typeof _satellite) && ("function" === typeof _satellite.pageBottom)) {
_satellite.pageBottom();
}
</script>