Don't use multiviews on docs.a.c/ansible/
We aren't using MultiViews for anything in the /ansible/ docs (it's
commonly used for translated docs but we're doing something different for
ansible.  Not sure about tower so this change will not affect
/ansible-tower/.)

We need to turn off MultiViews because it interferes with redirects that
we are generating.  For example, we have a redirect from
2.10/plugins/become/runas.html =>
  2.10/collections/ansible/builtin/runas_become.html
With MultiViews on, apache looks for the directory 2.10/plugins/become,
doesn't find it, but does find 2.10/plugins/become.html.  It decides to
use that html file instead of the directory and then it prevents the
redirect from working properly.

If we have to turn MultiViews on in the future, another workaround would
be to create empty directories in the old locations (like:
mkdir 2.10/plugins/become/ )

(cherry picked from commit 8c1c00be856114c395fb630c54346651191edaae)